fix open directory, see other notes below
if an image cant load correctly for previews. do not load it. right now it shows a broken image symbol. Just don't show it.
on the versions tab, change the "set path" button to "Change". when this button is clicked open the current path
when the executable "Change" button is pressed, open the executable folder.
Change "short name" on the records tab to "Mappings". this box should list all of the mapper names separated by a comma